changed double to int

+++ lib/models/room.dart
@@ -1,7 +1,7 @@
 class Room {
   final String id;
   final String name;
-  final double deviceCount;
+  final int deviceCount;
 
   const Room({
     required this.id,
